requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
atlassian-jwt==3.0.0
atlassian-python-api==3.28.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-redis==5.2.0
environ==1.0
gunicorn==20.1.0
hiredis==2.0.0
idna==3.4
jira==2.0.0
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
pbr==5.10.0
psycopg==3.1.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-jwt==0.5.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
